backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.103.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# File Processing
python-magic>=0.4.27
PyMuPDF>=1.22.0
python-docx>=0.8.11
pdfplumber>=0.10.0

# ML and Text Processing
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
langchain>=0.0.200
sentence-transformers>=2.2.2
chromadb>=0.4.0
openai>=0.28.0

# Utilities
python-dotenv>=1.0.0
pydantic[email]>=2.3.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
email-validator>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
pytest-cov>=4.1.0
faker>=19.3.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0
sentry-sdk>=1.29.0